+++ src/shared/ExtensionMessage.ts
@@ -24,11 +24,12 @@ export interface ExtensionMessage {
 		| "haiTaskData"
 		| "haiAction"
 		| "haiConfig"
-		| "apiKeyValidation"
-		| "embeddingValidation"
+		| "llmConfigValidation"
+		| "embeddingConfigValidation"
 		| "mcpServers"
 		| "existingFiles"
 	text?: string
+	bool?: boolean
 	action?:
 		| "chatButtonClicked"
 		| "mcpButtonClicked"